Created a build system and modularized JavaScript
Added a build system to organize the sources into an organized and
deployable file tree. The output of the build process is a system ready
to be served.

The addition of a build system also brings the possibility of
modularization of the JavaScript code. With the use of Rollup to
assemble the multiple files is possible to split the code into modules.
This commit apply the necessary changes to conform current js files
with the es6 modules standard.