Changed the JavaScript design of the components
The architecture of UI components was very messy and unclean, some
components implemented the functionality os all corresponding elements
on the page in a single object, others implemented constructors for each
element, some with methods declared on prototype others on each
instance. The interface of each component wasn't clear either, some of
them didn't allow straight access to the methods or the respective HTML
elements.

Now every component is implemented in the same standardized way: they
inherit common properties from the component constructor and define a
constructor that takes the HTML element represented by that component
as a parameter. To initialize the components for the entire page and
make them accessible the index calls the constructor of each component
for all corresponding elements in the page and stores the created
objects exposing them through a function that match selectors like
querySelectorAll.